pyproject.toml
[project]
name = "ergodicts"
version = "0.1.0"
description = "Ergodic TS — tools for time-series forecasting"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "snowflake-connector-python[pandas]",
    "python-dotenv",
    "pydantic>=2.0",
]

[project.optional-dependencies]
viz = ["graphviz"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ergodicts"]

[dependency-groups]
dev = [
    "ipykernel>=7.2.0",
    "pytest>=8.0",
]
docs = [
    "mkdocs-material",
    "mkdocstrings[python]",
]
